Folder refactoring and fixed various bugs
