+++ new-experiments/activity_change_recognition/unsupervised/embeddings/activity_change_recognition_context_emb_sim.py
@@ -1,30 +1,22 @@
 import json
 import sys
-import os
 import numpy as np
 import pandas as pd
-import ruptures as rpt
-import itertools
-import operator
-import csv
 import argparse
 
-from gensim.models import Word2Vec
 from sklearn import preprocessing
-from keras.preprocessing.text import Tokenizer
-from scipy import spatial
-from scipy.stats import entropy, norm
-from datetime import datetime as dt
-
 from pylab import *
-from scipy import linalg
-from scipy.stats import norm
-
 from math import sqrt
+
+from utils.activity_change_preprocessing import *
+from utils.activity_change_save_results import *
+from utils.activity_change_evaluation import *
+
+from embeddings_utils.create_embedding_matrix import *
+from context_similarity.calculate_context_similarity import *
 
 import multiprocessing
 from gensim.models import Word2Vec
-from gensim.models.word2vec import LineSentence
 
 # Kasteren dataset DIR
 DIR = '../../kasteren_house_a/'
@@ -34,188 +26,6 @@ DATASET_CSV = DIR + 'base_kasteren_reduced'
 UNIQUE_ACTIONS = DIR + 'unique_actions.json'
 # Context information for the actions in the dataset
 CONTEXT_OF_ACTIONS = DIR + 'context_model.json'
-
-def create_dirs(results_dir):
-    os.makedirs(results_dir, exist_ok=True)
-    os.makedirs(results_dir + "1s", exist_ok=True)
-    os.makedirs(results_dir + "5s", exist_ok=True)
-    os.makedirs(results_dir + "10s", exist_ok=True)
-    os.makedirs(results_dir + "detection_delays", exist_ok=True)
-    os.makedirs(results_dir + "word2vec_models", exist_ok=True)
-
-def get_unixtime(dt64):
-    return dt64.astype('datetime64[s]').astype('int')
-
-def convert_epoch_time_to_hour_of_day(epoch_time_in_seconds):
-    d = dt.fromtimestamp(epoch_time_in_seconds)
-    return d.strftime('%H')
-
-def convert_epoch_time_to_day_of_the_week(epoch_time_in_seconds):
-    d = dt.fromtimestamp(epoch_time_in_seconds)
-    return d.strftime('%A')
-
-def get_seconds_past_midnight_from_epoch(epoch_time_in_seconds):
-    date = dt.fromtimestamp(epoch_time_in_seconds)
-    seconds_past_midnight = (date - date.replace(hour=0, minute=0, second=0, microsecond=0)).total_seconds()
-    return seconds_past_midnight
-
-def create_action_embedding_matrix(tokenizer, model, embedding_size):  
-    action_index = tokenizer.word_index
-    embedding_matrix = np.zeros((len(action_index) + 1, embedding_size))
-    unknown_actions = {}    
-    for action, i in list(action_index.items()):
-        try:            
-            embedding_vector = model[action]
-            embedding_matrix[i] = embedding_vector            
-        except:
-            if action in unknown_actions:
-                unknown_actions[action] += 1
-            else:
-                unknown_actions[action] = 1
-    
-    return embedding_matrix, unknown_actions
-
-def calculate_context_similarity(actions, embedding_action_matrix, position, window):
-    target_action_vector = embedding_action_matrix[actions[position]]
-    target_action_vector_context_sim = 0.0
-    counter = window * 2
-    for i in range(1, window+1):
-        # right context we search for similarity
-        if position+i < len(actions):
-            right_sim = 1 - spatial.distance.cosine(target_action_vector, embedding_action_matrix[actions[position+i]])
-            target_action_vector_context_sim += right_sim
-        else:
-            counter -= 1
-        # left context we search for disimilarity (1 - similarity)
-        if position-i >= 0:
-            left_sim = 1 - max(0, 1 - spatial.distance.cosine(target_action_vector, embedding_action_matrix[actions[position-i]]))
-            target_action_vector_context_sim += left_sim
-        else:
-            counter -= 1
-    target_action_vector_context_sim = target_action_vector_context_sim / counter
-    return target_action_vector_context_sim
-
-def prepare_x_y_activity_change(df):
-    actions = df['action'].values
-    dates = list(df.index.values)
-    timestamps = list(map(get_unixtime, dates))
-    hours = list(map(convert_epoch_time_to_hour_of_day, timestamps))
-    days = list(map(convert_epoch_time_to_day_of_the_week, timestamps))
-    seconds_past_midnight = list(map(get_seconds_past_midnight_from_epoch, timestamps))
-    activities = df['activity'].values
-
-    tokenizer_action = Tokenizer(lower=False)
-    tokenizer_action.fit_on_texts(actions.tolist())
-    action_index = tokenizer_action.word_index
-    
-    actions_by_index = []
-    for action in actions:
-        actions_by_index.append(action_index[action])
-    
-    X = []
-    y = []
-    last_activity = None
-    for i in range(0, len(actions)):
-        X.append(actions_by_index[i])
-        if (i == 0):
-            y.append(0)
-        elif last_activity == activities[i]:
-            y.append(0)
-        else:
-            y.append(1)
-        last_activity = activities[i]
-    
-    X = np.array(X)
-    timestamps = np.array(timestamps)
-    days = np.array(days)
-    hours = np.array(hours)
-    seconds_past_midnight = np.array(seconds_past_midnight)
-    y = np.array(y)
-    
-    return X, timestamps, days, hours, seconds_past_midnight, y, tokenizer_action
-
-def check_detected_change_point_with_offset(timestamps, y, position, offset):
-    timestamp_ref = timestamps[position]
-
-    # right
-    for i in range(position, len(y)):
-        difference = timestamps[i] - timestamp_ref
-        if offset < difference:
-            break
-        else:
-            if y[i] == 1:
-                return True
-    # left
-    i = position - 1
-    while i >= 0:
-        difference = timestamp_ref - timestamps[i]
-        if offset < difference:
-            break
-        else:
-            if y[i] == 1:
-                return True
-        i -= 1
-
-    return False
-
-def get_conf_matrix_with_offset_strategy(similarities, y, timestamps, max_context_distance, offset):
-    cf_matrix = np.zeros((2,2))
-
-    for i in range(0, len(similarities)):
-        if similarities[i] > max_context_distance:
-            correctly_detected = check_detected_change_point_with_offset(timestamps, y, i, offset)
-            if correctly_detected:
-                cf_matrix[1][1] += 1
-            else:
-                cf_matrix[0][1] += 1
-        else:
-            if y[i] == 0:
-                cf_matrix[0][0] += 1
-            else:
-                cf_matrix[1][0] += 1
-    
-    return cf_matrix
-
-def search_nearest_change_point(y, timestamps, position):
-    timestamp_ref = timestamps[position]
-
-    # right
-    right_nearest_cp_difference = sys.maxsize
-    for i in range(position, len(y)):
-        if y[i] == 1:
-            right_nearest_cp_difference = timestamps[i] - timestamp_ref
-            break
-    
-    # left
-    left_nearest_cp_difference = sys.maxsize
-    i = position - 1
-    while i >= 0:
-        if y[i] == 1:
-            left_nearest_cp_difference = timestamp_ref - timestamps[i]
-            break
-        i -= 1
-
-    return min(right_nearest_cp_difference, left_nearest_cp_difference)
-
-def get_detection_delay(similarities, y, timestamps, max_context_distance):
-    detection_delay = 0
-
-    counter = 0
-    for i in range(0, len(similarities)):
-        if similarities[i] > max_context_distance:
-            counter += 1
-            detection_delay += search_nearest_change_point(y, timestamps, i)
-    
-    if counter == 0: 
-        return 0
-    else:
-        return detection_delay / counter
-
-def save_pop_results_to_file(results_dir, results, seconds_for_detection, max_context_distance_num):
-    np.savetxt(results_dir + str(seconds_for_detection) + 's/results_' + str(seconds_for_detection) + '_' + str(max_context_distance_num) + '_TPR' + '.csv', results[0][max_context_distance_num])
-    np.savetxt(results_dir + str(seconds_for_detection) + 's/results_' + str(seconds_for_detection) + '_' + str(max_context_distance_num) + '_TNR' + '.csv', results[1][max_context_distance_num])
-    np.savetxt(results_dir + str(seconds_for_detection) + 's/results_' + str(seconds_for_detection) + '_' + str(max_context_distance_num) + '_FPR' + '.csv', results[2][max_context_distance_num])
-    np.savetxt(results_dir + str(seconds_for_detection) + 's/results_' + str(seconds_for_detection) + '_' + str(max_context_distance_num) + '_G-MEAN' + '.csv', results[3][max_context_distance_num])
 
 def main(argv):
     np.set_printoptions(threshold=sys.maxsize)
@@ -305,7 +115,7 @@ def main(argv):
     embedding_size = args.embedding_size
     RESULTS_DIR = "/" + args.results_dir + "/" + args.results_folder + "/context_window_" + str(context_window_size) + "_window_" + str(window_size) + "_iterations_" + str(iterations) + "_embedding_size_" + str(embedding_size) + "/" + args.train_or_test + "/"
     # create dirs for saving results
-    create_dirs(RESULTS_DIR)
+    create_dirs(RESULTS_DIR, word2vec=True)
     # check actions input shape
     print("Input action shape: " + str(X.shape))
     # repeat exe iterations
@@ -323,12 +133,13 @@ def main(argv):
             models.append(model)
         # if test set, load word2vec model
         elif args.train_or_test == 'test':
-            model = Word2Vec.load(RESULTS_DIR + 'word2vec_models/' + str(exe) + '_execution.model') #TODO check this
+            model = Word2Vec.load(RESULTS_DIR.replace("test", "train") + 'word2vec_models/' + str(e) + '_execution.model')
         # create embedding matrix
         embedding_action_matrix, unknown_actions = create_action_embedding_matrix(tokenizer_action, model, embedding_size)
-        # calculate similarities using word2vec embeddings
+        # calculate context similarities using word2vec embeddings
         similarities = []
-        for i in range(0, len(X)-1):
+        similarities.append(0)
+        for i in range(1, len(X)):
             context_similarity = calculate_context_similarity(X, embedding_action_matrix, i, context_window_size)
             similarities.append(context_similarity)
         # prepare change detection with offset using different max_context_distance values
