+++ new-experiments/activity_change_recognition/unsupervised/embeddings/embeddings_utils/create_embedding_matrix.py
@@ -1,0 +1,17 @@
+import numpy as np
+
+def create_action_embedding_matrix(tokenizer, model, embedding_size):  
+    action_index = tokenizer.word_index
+    embedding_matrix = np.zeros((len(action_index) + 1, embedding_size))
+    unknown_actions = {}    
+    for action, i in list(action_index.items()):
+        try:            
+            embedding_vector = model[action]
+            embedding_matrix[i] = embedding_vector            
+        except:
+            if action in unknown_actions:
+                unknown_actions[action] += 1
+            else:
+                unknown_actions[action] = 1
+    
+    return embedding_matrix, unknown_actions
