+++ new-experiments/activity_change_recognition/unsupervised/likelihood/RuLSIF/feature_extraction_RuLSIF.py
@@ -1,0 +1,115 @@
+import numpy as np
+import itertools
+import operator
+
+from scipy.stats import entropy
+
+##################################################################################################################
+# Feature extraction approach of Real-Time Change Point Detection with Application to Smart Home Time Series Data
+# https://ieeexplore.ieee.org/document/8395405
+# START
+##################################################################################################################
+
+def most_common(L):
+  SL = sorted((x, i) for i, x in enumerate(L))
+
+  groups = itertools.groupby(SL, key=operator.itemgetter(0))
+
+  def _auxfun(g):
+    item, iterable = g
+    count = 0
+    min_index = len(L)
+    for _, where in iterable:
+      count += 1
+      min_index = min(min_index, where)
+    return count, -min_index
+
+  return max(groups, key=_auxfun)[0]
+
+def calc_entropy(labels, base=None):
+  value,counts = np.unique(labels, return_counts=True)
+  return entropy(counts, base=base)
+
+def day_to_int(day):
+    switcher = {"Monday": 0, "Tuesday": 1, "Wednesday": 2, "Thursday": 3, "Friday": 4, "Saturday": 5, "Sunday": 6}
+    return switcher[day]
+
+def location_to_int(location):
+    switcher = {"Bathroom": 0, "Kitchen": 1, "Bedroom": 2, "Hall": 3}
+    return switcher[location]
+
+def extract_actions_from_window(actions, timestamps, position, window_size=30):
+    actions_from_window = []
+    timestamps_from_window = []
+
+    for i in range(position-(window_size), position):
+        if i >= 0:
+            actions_from_window.append(actions[i])
+            timestamps_from_window.append(timestamps[i])
+    actions_from_window.append(actions[position])
+    timestamps_from_window.append(timestamps[position])
+
+    return actions_from_window, timestamps_from_window
+
+def extract_features_from_window(actions, previous_actions_1, previous_actions_2, locations, timestamps):
+    features_from_window = []
+    
+    most_recent_sensor = actions[len(actions)-1]
+    first_sensor_in_window = actions[0]
+    window_duration = timestamps[len(actions)-1] - timestamps[0]
+    if previous_actions_1 is not None:
+        most_frequent_sensor_1 = most_common(previous_actions_1)
+    else:
+        most_frequent_sensor_1 = 0 # non-existing sensor index
+    if previous_actions_2 is not None:
+        most_frequent_sensor_2 = most_common(previous_actions_2)
+    else:
+        most_frequent_sensor_2 = 0 # non-existing sensor index
+    last_sensor_location = location_to_int(locations[actions[len(actions)-1]])
+    # last_motion_sensor_location -> no aplica en kasteren !!!!!
+    entropy_based_data_complexity = calc_entropy(actions)
+    time_elapsed_since_last_sensor_event = timestamps[len(actions)-1] - timestamps[len(actions)-2]
+
+    features_from_window.append(most_recent_sensor)
+    features_from_window.append(first_sensor_in_window)
+    features_from_window.append(window_duration)
+    features_from_window.append(most_frequent_sensor_1)
+    features_from_window.append(most_frequent_sensor_2)
+    features_from_window.append(last_sensor_location)
+    features_from_window.append(entropy_based_data_complexity)
+    features_from_window.append(time_elapsed_since_last_sensor_event)
+
+    return features_from_window
+
+def extract_features_from_sensors(actions, unique_actions, all_actions, position, timestamps, all_timestamps):
+    features_from_sensors = []
+    
+    # count of events for each sensor in window
+    for action in unique_actions:
+        counter = 0
+        for action_fired in actions:
+            if action == action_fired:
+                counter += 1
+        features_from_sensors.append(counter)
+    
+    # elapsed time for each sensor since last event
+    found_actions = []
+    counter = position
+    for action in unique_actions:
+        while(counter > 0):
+            if action == all_actions[counter]:
+                found_actions.append(action)
+                features_from_sensors.append(timestamps[len(timestamps)-1] - all_timestamps[counter])
+                break
+            counter -= 1
+        if action not in found_actions:
+            features_from_sensors.append(all_timestamps[len(all_timestamps)-1] - all_timestamps[0]) # maximun time possible
+        counter = position
+
+    return features_from_sensors
+
+##################################################################################################################
+# Feature extraction approach of Real-Time Change Point Detection with Application to Smart Home Time Series Data
+# https://ieeexplore.ieee.org/document/8395405
+# END
+##################################################################################################################
