+++ run_container.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 
-sudo docker stop ubermejo-activity-recognition-3
-sudo docker rm ubermejo-activity-recognition-3
-sudo docker run -it --ipc=host --gpus "device=0" --name ubermejo-activity-recognition-3 -v activity-recognition:/results ubermejo/activity_recognition bin/bash
+sudo docker stop ubermejo-activity-recognition
+sudo docker rm ubermejo-activity-recognition
+sudo docker run -it --ipc=host --gpus "device=0" --name ubermejo-activity-recognition -v activity-recognition:/results ubermejo/activity_recognition bin/bash
